fix an issue with non-file visiting buffers
In some cases, there we can have temporary/special buffers with `python-mode`
that doesn't visit a file (`buffer-file-name` is nil). In these cases, the
current `pet-mode` gets called and cause an error.

This has been observed in `mmm-mode` when `python-mode` is used as a sub-mode,
although, it can happen in other situations.