pipenv: Ignore additional pipenv output
Pipenv can print additional output that would end up in our variables, e. g., if
a `.env` file exists, `python-shell-virtualenv-root` ends up as `Loading .env
environment variables...\n/home/user/.local/share/virtualenvs/project`.

This change is to:

a) call pipenv with `--quiet`, so that it does not print extra information in
   the first place;

b) pass `'(t nil)` as `buffer` argument to `process-file`, so that it ignores
   stderr and the buffer only contains stdout.

Signed-off-by: Thomas Schneider <qsx@chaotikum.eu>